refactor(activity): extract transaction query hooks
